forgot to call update on pSystem
there are a lot of "hardcoded" values right now; artifacts start to
disappear as the first three values in mesh's constructor (in main)
increase since there are more cubes, but it gets SLOW (for me
especially). building the mesh is also slow, though it only gets built
in the very beginning. no support for multiple fluids, and I have no
idea how to give it that